+++ src/index.ts
@@ -3,3 +3,4 @@ export { MongodbRepo, repoProvider, IRepoRegistry } from './repository';
 export { transaction, repoSession } from './decorators';
 export { IMongodbConfig, ObjectId, MongoId } from './types';
 export { mongodbConfigSchema } from './config';
+export * from './utils';